+++ src/Button.tsx
@@ -1,6 +1,6 @@
 /** @jsx createElement */
 import { createElement, FC } from 'react';
-import { getModifierClass } from './utils/getModifierClass';
+import getModifierClass from './utils/getModifierClass';
 
 type ButtonElmProps = JSX.IntrinsicElements['button'];